+++ gulp/metalsmith.js
@@ -22,6 +22,8 @@ const permalinks = require("metalsmith-permalinks");
 const sitemap = require("metalsmith-mapsite");
 const debug = require("metalsmith-debug");
 const discoverPartials = require("metalsmith-discover-partials");
+const collections = require("metalsmith-collections");
+const drafts = require("metalsmith-drafts");
 const gravityAssets = require("./gravity-assets.js");
 const jobListings = require("./metalsmith-job-listings.js");
 const flourishShapes = require("../src/flourishes/shapes.json");
@@ -55,8 +57,15 @@ function metalsmith() {
     }
     return null;
   });
-  handlebars.registerHelper("flourishShapes", function(name) {
-    const template = flourishTemplate(flourishShapes[name]);
+  handlebars.registerHelper("flourishShapes", function(name, excludeGradient) {
+    const contextData = flourishShapes[name];
+    // If no 2nd arg is given to the helper, Handlebars passed in its options
+    // object here
+    if (typeof excludeGradient === "object") {
+      excludeGradient = false;
+    }
+    contextData.excludeGradient = excludeGradient;
+    const template = flourishTemplate(contextData);
     return new handlebars.SafeString(template);
   });
 
@@ -84,7 +93,8 @@ function metalsmith() {
           .metadata({
             site: {
               title: config.title,
-              url: siteEnv.url
+              url: siteEnv.url,
+              googleAnalyticsTrackingId: siteEnv.googleAnalyticsTrackingId
             },
             build: {
               excludeRobots: siteEnv.excludeRobots
@@ -92,8 +102,8 @@ function metalsmith() {
 
             // Defaults for Twitter card meta tags
             // (See: https://developer.twitter.com/en/docs/tweets/optimize-with-cards/guides/getting-started)
-            twitterCard: "summary",
-            twitterSite: "@Buildit_tech",
+            twitterCard: "summary_large_image",
+            twitterSiteId: "910173909592485889", // ID for @Buildit_tech
 
             // Defaults for OpenGraph Protocol (OGP) meta tags
             // (See: http://ogp.me/)
@@ -102,6 +112,12 @@ function metalsmith() {
             ogImageAlt: "buildit @ wipro digital"
           })
           .use(jobListings())
+          .use(
+            collections({
+              locations: `locations/*.md`
+            })
+          )
+          .use(drafts())
           .use(pageTitles())
           .use(markdown())
           .use(