+++ CHANGELOG.md
@@ -15,8 +15,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Animated homepage hero
 - Client-side JavaScript transpilation and bundling (using Babel and Rollup)
 - JavaScript unit tests using Jest
+- [Cypress](https://www.cypress.io/) integration tests
+- [Pa11y](http://pa11y.org/) accessibility tests
 - ESLint with Prettier
 - Templated `robots.txt` that can be configured per deployment environment
+- Image optimisation task
 
 ### Changed
 - Site build now uses Metalsmith